requirements.txt
canonicalwebteam.flask-base==2.1.0
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.search==2.1.1
canonicalwebteam.image-template==1.5.0
mistune==0.8.4
pyyaml==6.0.2

# Observability
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-wsgi==0.53b1
opentelemetry-instrumentation-flask==0.53b1
opentelemetry-instrumentation-requests==0.53b1
opentelemetry-sdk==1.32.1
